+++ src/styles/index.js
@@ -1,4 +1,5 @@
 import variants from './variants';
+import responsive from './responsive';
 import * as keyframes from './keyframes';
 
-export {variants, keyframes};
+export {responsive, variants, keyframes};